'Tweak' to fix transparency of checkboxes and radio items for Ubuntu14&15
Transparency of checkboxes and radio items in themes is pretty common when trying to create a unified theme for Ubuntu 14.xx and also 15.xx, the most logical thing to do is create two versions of a theme for each version of Ubuntu. But the easiest trick I've tried for fixing this, is creating a "gtk-widgets-assets.css" script with the two types of code in one file, that way Ubuntu 15.xx will read what it can and Ubuntu 14.xx will read lines with the previous format. Until now I haven't found any bugs with doing this and checkboxes and radio items work now for the two versions with one single theme. I submit this for review.